+++ backend/pyproject.toml
@@ -11,6 +11,7 @@ dependencies = [
     "requests>=2.32.3",
     "openai>=1.55.0",
     "uvicorn>=0.32.0",
+    "httpx>=0.27.2",
 ]
 
 [dependency-groups]
